docs: document buildQueryExpression (#1403)
https://coveord.atlassian.net/browse/KIT-1131
