--- packages/headless/src/utils/expression-builder/common/negatable.ts
@@ -1,7 +1,0 @@
-export interface Negatable {
-  negate?: boolean;
-}
-
-export function getNegationPrefix(config: Negatable) {
-  return config.negate ? 'NOT ' : '';
-}
