--- packages/headless/src/utils/expression-builder/common/part.ts
@@ -1,3 +1,0 @@
-export interface Part {
-  toString(): string;
-}
