--- packages/headless/src/utils/expression-builder/numeric-range-field/numeric-range-field.ts
@@ -1,22 +1,0 @@
-import {getNegationPrefix, Negatable} from '../common/negatable';
-import {getOperatorSymbol} from '../common/operator';
-import {Part} from '../common/part';
-
-export interface NumericRangeFieldExpression extends Negatable {
-  field: string;
-  from: number;
-  to: number;
-}
-
-export function buildNumericRangeField(
-  config: NumericRangeFieldExpression
-): Part {
-  return {
-    toString() {
-      const prefix = getNegationPrefix(config);
-      const {field, from, to} = config;
-      const operator = getOperatorSymbol('isExactly');
-      return `${prefix}@${field}${operator}${from}..${to}`;
-    },
-  };
-}
