--- packages/headless/src/utils/expression-builder/query-extension/query-extension.test.ts
@@ -1,47 +1,0 @@
-import {createExpressionBuilder} from '../expression-builder';
-import {buildQueryExtension} from './query-extension';
-
-describe('#buildQueryExtension', () => {
-  describe('#toString', () => {
-    it('with no parameters', () => {
-      const builder = buildQueryExtension({
-        name: 'q',
-        parameters: [],
-      });
-
-      expect(builder.toString()).toBe('$q()');
-    });
-
-    it('with multiple parameters', () => {
-      const fieldExpression = createExpressionBuilder({
-        delimiter: 'or',
-      }).addStringField({
-        field: 'documenttype',
-        operator: 'isExactly',
-        values: ['Book'],
-      });
-
-      const modifierExpression = createExpressionBuilder({
-        delimiter: 'or',
-      }).addKeyword({expression: '100'});
-
-      const builder = buildQueryExtension({
-        name: 'qre',
-        parameters: [
-          {
-            name: 'expression',
-            value: fieldExpression,
-          },
-          {
-            name: 'modifier',
-            value: modifierExpression,
-          },
-        ],
-      });
-
-      expect(builder.toString()).toBe(
-        '$qre(expression: @documenttype=="Book", modifier: 100)'
-      );
-    });
-  });
-});
