+++ packages/headless/src/utils/query-expression/common/negatable.ts
@@ -1,0 +1,10 @@
+export interface Negatable {
+  /**
+   * If `true`, the inverse expression will be created.
+   */
+  negate?: boolean;
+}
+
+export function getNegationPrefix(config: Negatable) {
+  return config.negate ? 'NOT ' : '';
+}
