+++ packages/headless/src/utils/query-expression/common/part.ts
@@ -1,0 +1,3 @@
+export interface Part {
+  toQuerySyntax(): string;
+}
