+++ packages/headless/src/utils/query-expression/numeric-range-field/numeric-range-field.test.ts
@@ -1,0 +1,26 @@
+import {buildNumericRangeField} from './numeric-range-field';
+
+describe('#buildNumericRangeField', () => {
+  describe('#toQuerySyntax', () => {
+    it('with #negate not specified', () => {
+      const builder = buildNumericRangeField({
+        field: 'size',
+        from: 10,
+        to: 20,
+      });
+
+      expect(builder.toQuerySyntax()).toBe('@size==10..20');
+    });
+
+    it('with #negate set to true', () => {
+      const builder = buildNumericRangeField({
+        field: 'size',
+        from: 10,
+        to: 20,
+        negate: true,
+      });
+
+      expect(builder.toQuerySyntax()).toBe('NOT @size==10..20');
+    });
+  });
+});
